Split up the GithubEngine class might make it easier to test?
