+++ NuKeeper/ContainerRegistration.cs
@@ -40,6 +40,7 @@ namespace NuKeeper
             container.Register<IRepositoryScanner, RepositoryScanner>();
 
             container.Register<GithubEngine>();
+            container.Register<IGithubRepositoryEngine, GithubRepositoryEngine>();
             container.Register<IRepositoryUpdater, RepositoryUpdater>();
             container.Register<IPackageUpdater, PackageUpdater>();
             container.Register<IForkFinder, ForkFinder>();
